Introduce new PUT method to update thoth config
